dc: hide 'going to remove...' message when disabling dc services
