+++ internal/cli/klog.go
@@ -8,52 +8,13 @@ import (
 	"os"
 	"regexp"
 
+	"github.com/tilt-dev/tilt/internal/filteredwriter"
+
 	"k8s.io/klog/v2"
 )
 
 // https://kubernetes.io/docs/reference/kubectl/cheatsheet/#kubectl-output-verbosity-and-debugging
 var klogLevel = 0
-
-type filteredWriter struct {
-	underlying io.Writer
-	filterFunc func(s string) bool
-	leftover   []byte
-}
-
-func (fw *filteredWriter) Write(buf []byte) (int, error) {
-	buf = append(fw.leftover, buf...)
-	start := 0
-	written := 0
-	for i, b := range buf {
-		if b == '\n' {
-			end := i
-			if buf[i-1] == '\r' {
-				end--
-			}
-			s := string(buf[start:end])
-
-			if !fw.filterFunc(s) {
-				n, err := fw.underlying.Write(buf[start : i+1])
-				written += n
-				if err != nil {
-					fw.leftover = append([]byte{}, buf[i+1:]...)
-					return written, err
-				}
-			}
-
-			start = i + 1
-		}
-	}
-
-	fw.leftover = append([]byte{}, buf[start:]...)
-
-	return written, nil
-}
-
-// lines matching `filterFunc` will not be output to the underlying writer
-func newFilteredWriter(underlying io.Writer, filterFunc func(s string) bool) io.Writer {
-	return &filteredWriter{underlying: underlying, filterFunc: filterFunc}
-}
 
 // everything on google indicates this warning is useless and should be ignored
 // https://github.com/kubernetes/kubernetes/issues/22024
@@ -81,7 +42,7 @@ func initKlog(w io.Writer) {
 	var tmpFlagSet = flag.NewFlagSet(os.Args[0], flag.ExitOnError)
 	klog.InitFlags(tmpFlagSet)
 	if klogLevel == 0 {
-		w = newFilteredWriter(w, filterMux(
+		w = filteredwriter.New(w, filterMux(
 			isResourceVersionTooOldRegexp.MatchString,
 			isGroupVersionEmptyRegexp.MatchString,
 		))
